API/pyproject.toml
[tool.poetry]
name = "aequivalent-api"
version = "0.1.0"
description = ""
authors = ["loichu <loic@humbert-web.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-phonenumber-field = "^6.1.0"
phonenumbers = "^8.12.47"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
